fix: 모임카드 타입 수정 및 폴더 이동
