+++ src/types/moim.ts
@@ -25,10 +25,39 @@ export interface Moim {
 }
 
 export interface MoimPayload {
+  id: number;
   title: string;
+  category: string;
   content: string;
   moim: string;
   place: string;
   date: string;
-  roles: Record<string, number>;
+  roles: {
+    prop1: number;
+    prop2: number;
+    prop3: number;
+  };
+}
+
+export interface moimCard {
+  id: number;
+  title: string;
+  category: string;
+  is_closed: boolean;
+  is_applied: boolean;
+  is_writer: boolean;
+  is_liked: boolean;
+  date: string;
+  place_name: string;
+  address: string;
+  latitude: number;
+  longitude: number;
+  max_people: number;
+  status: number;
+  roles: {
+    prop1: number;
+    prop2: number;
+    prop3: number;
+  };
+  img_url?: string;
 }
